+++ emu/tr_utils.c
@@ -183,7 +183,7 @@ BTNptr trie_lookup(int nTerms, CPtr termVector, BTNptr trieRoot,
       int i;
 
       varArray[0] = tstTrail.top - tstTrail.base;
-      for ( i = 1; i <= varArray[0]; i++ )
+      for ( i = 1; (byte)i <= varArray[0]; i++ )
 	varArray[i] = (Cell)tstTrail.base[i - 1];
     }
 
@@ -240,7 +240,7 @@ Cell build_ret_term(int arity, Cell termVector[]) {
     return makestring(ret_psc[0]);  /* return as a term */
   else {
     ret_term = hreg;  /* pointer to where ret(..) will be built */
-    sym = insert("ret", arity, (Psc)flags[CURRENT_MODULE], &new);
+    sym = insert("ret", (byte)arity, (Psc)flags[CURRENT_MODULE], &new);
     new_heap_functor(hreg, pair_psc(sym));
     for ( i = 0; i < arity; i++ )
       nbldval(termVector[i]);
@@ -829,7 +829,7 @@ void breg_retskel(void)
       hreg = sreg;
 */
       bind_cs((CPtr)ptoc_tag(3), hreg);
-      sym = insert("ret", Nvars, (Psc)flags[CURRENT_MODULE], &new);
+      sym = insert("ret", (byte)Nvars, (Psc)flags[CURRENT_MODULE], &new);
       new_heap_functor(hreg, sym->psc_ptr);
 #ifdef CHAT
       for (i = Nvars; i > 0; i--) {
